Consolidate responses continuation tests
